Better way to determine seeking for youtube
Instead of the catchRougePauseEvent code, check check the timedelta
in the pause event.  Since youtube fires the pause event on seek, this
seems like the best place to check it.

This method avoids issues where the CPU load is too high, so the
monitorCurrentTime() timeout gets delayed and seeking/seek events get
emitted when they shouldn't.